requirements.txt
pandas==1.5.3
praw==7.7.0
python-dotenv==1.0.0
requests==2.22.0
